Support LWS-based multi node with data parallelism of MoE models (#716)
* Support LWS-based multi node with data parallelism of MoE models

For more details, please read:
- https://docs.google.com/document/d/1mSYsWQEbp4Oq50ghFWUun7OgagbuUJivAC8Ds-pu-xU/edit?tab=t.0
- https://docs.google.com/presentation/d/1_nSxtf0h1lO3EySk5FEP9Neqym9h3VYGRsMt_ZnRaGw/edit?slide=id.p#slide=id.p
- https://docs.google.com/document/d/19dAjS6gqYWRJ0FQMYCwG-XJHIgWwmvsS-pBynwPHhwo/edit?tab=t.0

Ref: RHOAIENG-28170


* Fix InferencePool selector for multi node and address reviews

- InferencePool selector now takes into account the multi node case
  sends requests to either leader or worker with index 0
- Addressed review comments
- Remove GPU presets as they are not generic enough to work with
  multi node cases and we can re-think the approach as future
  iteration
- Add LWS to scheme and watch


---------

+++ pkg/controller/llmisvc/config_merge_test.go
@@ -293,7 +293,7 @@ func TestMergeSpecs(t *testing.T) {
 				{
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Tensor: ptr.To[int64](2),
+							Tensor: ptr.To[int32](2),
 						},
 					},
 				},
@@ -301,7 +301,7 @@ func TestMergeSpecs(t *testing.T) {
 				{
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Pipeline: ptr.To[int64](4),
+							Pipeline: ptr.To[int32](4),
 						},
 					},
 				},
@@ -310,8 +310,8 @@ func TestMergeSpecs(t *testing.T) {
 				WorkloadSpec: v1alpha1.WorkloadSpec{
 					// Both parallelism values should be present
 					Parallelism: &v1alpha1.ParallelismSpec{
-						Tensor:   ptr.To[int64](2),
-						Pipeline: ptr.To[int64](4),
+						Tensor:   ptr.To[int32](2),
+						Pipeline: ptr.To[int32](4),
 					},
 				},
 			},
@@ -385,8 +385,8 @@ func TestMergeSpecs(t *testing.T) {
 				{
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Tensor:   ptr.To[int64](1),
-							Pipeline: ptr.To[int64](1),
+							Tensor:   ptr.To[int32](1),
+							Pipeline: ptr.To[int32](1),
 						},
 						Worker: &corev1.PodSpec{
 							Containers: []corev1.Container{
@@ -431,8 +431,8 @@ func TestMergeSpecs(t *testing.T) {
 				{
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Tensor:   ptr.To[int64](4),
-							Pipeline: ptr.To[int64](2),
+							Tensor:   ptr.To[int32](4),
+							Pipeline: ptr.To[int32](2),
 						},
 						Worker: &corev1.PodSpec{
 							Containers: []corev1.Container{
@@ -478,8 +478,8 @@ func TestMergeSpecs(t *testing.T) {
 				},
 				WorkloadSpec: v1alpha1.WorkloadSpec{
 					Parallelism: &v1alpha1.ParallelismSpec{
-						Tensor:   ptr.To[int64](4),
-						Pipeline: ptr.To[int64](2),
+						Tensor:   ptr.To[int32](4),
+						Pipeline: ptr.To[int32](2),
 					},
 					Worker: &corev1.PodSpec{
 						Containers: []corev1.Container{
@@ -538,8 +538,8 @@ func TestMergeSpecs(t *testing.T) {
 				{
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Tensor:   ptr.To[int64](1),
-							Pipeline: ptr.To[int64](1),
+							Tensor:   ptr.To[int32](1),
+							Pipeline: ptr.To[int32](1),
 						},
 						Worker: &corev1.PodSpec{
 							Containers: []corev1.Container{
@@ -570,8 +570,8 @@ func TestMergeSpecs(t *testing.T) {
 				{
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Tensor:   ptr.To[int64](4),
-							Pipeline: ptr.To[int64](2),
+							Tensor:   ptr.To[int32](4),
+							Pipeline: ptr.To[int32](2),
 						},
 						Worker: &corev1.PodSpec{
 							Containers: []corev1.Container{
@@ -620,8 +620,8 @@ func TestMergeSpecs(t *testing.T) {
 				},
 				WorkloadSpec: v1alpha1.WorkloadSpec{
 					Parallelism: &v1alpha1.ParallelismSpec{
-						Tensor:   ptr.To[int64](4),
-						Pipeline: ptr.To[int64](2),
+						Tensor:   ptr.To[int32](4),
+						Pipeline: ptr.To[int32](2),
 					},
 					Worker: &corev1.PodSpec{
 						Containers: []corev1.Container{
@@ -651,8 +651,8 @@ func TestMergeSpecs(t *testing.T) {
 				{
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Tensor:   ptr.To[int64](1),
-							Pipeline: ptr.To[int64](1),
+							Tensor:   ptr.To[int32](1),
+							Pipeline: ptr.To[int32](1),
 						},
 						Worker: &corev1.PodSpec{
 							Containers: []corev1.Container{
@@ -713,8 +713,8 @@ func TestMergeSpecs(t *testing.T) {
 				},
 				WorkloadSpec: v1alpha1.WorkloadSpec{
 					Parallelism: &v1alpha1.ParallelismSpec{
-						Tensor:   ptr.To[int64](1),
-						Pipeline: ptr.To[int64](1),
+						Tensor:   ptr.To[int32](1),
+						Pipeline: ptr.To[int32](1),
 					},
 					Worker: &corev1.PodSpec{
 						Containers: []corev1.Container{
@@ -970,7 +970,7 @@ func TestMergeSpecs(t *testing.T) {
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Replicas: ptr.To[int32](1),
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Tensor: ptr.To[int64](2),
+							Tensor: ptr.To[int32](2),
 						},
 					},
 					Router: &v1alpha1.RouterSpec{
@@ -998,7 +998,7 @@ func TestMergeSpecs(t *testing.T) {
 					WorkloadSpec: v1alpha1.WorkloadSpec{
 						Replicas: ptr.To[int32](5),
 						Parallelism: &v1alpha1.ParallelismSpec{
-							Pipeline: ptr.To[int64](4),
+							Pipeline: ptr.To[int32](4),
 						},
 					},
 					Router: &v1alpha1.RouterSpec{
@@ -1031,8 +1031,8 @@ func TestMergeSpecs(t *testing.T) {
 				WorkloadSpec: v1alpha1.WorkloadSpec{
 					Replicas: ptr.To[int32](5),
 					Parallelism: &v1alpha1.ParallelismSpec{
-						Tensor:   ptr.To[int64](2), // Base tensor preserved
-						Pipeline: ptr.To[int64](4), // Override pipeline
+						Tensor:   ptr.To[int32](2), // Base tensor preserved
+						Pipeline: ptr.To[int32](4), // Override pipeline
 					},
 				},
 				Router: &v1alpha1.RouterSpec{
